# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MAY2026 Team 041 API
  version: 0.1.0
paths:
...
  /api/v1/members:
    get:
      tags:
      - members
      summary: List Members
      operationId: list_members_api_v1_members_get
      security:
      - HTTPBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Match against full name or email
          title: Search
        description: Match against full name or email
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Exact role name match
          title: Role
        description: Exact role name match
      - name: active_only
        in: query
        required: false
        schema:
          type: boolean
          description: Only include active accounts
          default: false
          title: Active Only
        description: Only include active accounts
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - members
      summary: Create Member
      operationId: create_member_api_v1_members_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/payments:
    post:
      tags:
      - payments
      summary: Create Payment
      operationId: create_payment_api_v1_payments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/admin/dashboard:
    get:
      tags:
      - admin
      summary: Get Dashboard
      operationId: get_dashboard_api_v1_admin_dashboard_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboardOut'
      security:
      - HTTPBearer: []
  /api/v1/admin/expenses:
    post:
      tags:
      - admin
      summary: Log Expense
      operationId: log_expense_api_v1_admin_expenses_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/audit-log:
    get:
      tags:
      - admin
      summary: Get Audit Log
      operationId: get_audit_log_api_v1_admin_audit_log_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditLogEntryOut'
                type: array
                title: Response Get Audit Log Api V1 Admin Audit Log Get
      security:
      - HTTPBearer: []
  /api/v1/admin/reports/revenue-by-plan:
    get:
      tags:
      - admin
      summary: Get Revenue By Plan
      operationId: get_revenue_by_plan_api_v1_admin_reports_revenue_by_plan_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueByPlanOut'
      security:
      - HTTPBearer: []
  /api/v1/admin/reports/profit-and-loss:
    get:
      tags:
      - admin
      summary: Get Profit And Loss
      operationId: get_profit_and_loss_api_v1_admin_reports_profit_and_loss_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfitAndLossOut'
      security:
      - HTTPBearer: []
  /api/v1/admin/reports/expense-breakdown:
    get:
      tags:
      - admin
      summary: Get Expense Breakdown
      operationId: get_expense_breakdown_api_v1_admin_reports_expense_breakdown_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseBreakdownOut'
      security:
      - HTTPBearer: []
  /api/v1/admin/reports/membership-growth:
    get:
      tags:
      - admin
      summary: Get Membership Growth
      operationId: get_membership_growth_api_v1_admin_reports_membership_growth_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipGrowthOut'
      security:
      - HTTPBearer: []
  /api/v1/admin/announcements:
    post:
      tags:
      - admin
      summary: Send Announcement
      operationId: send_announcement_api_v1_admin_announcements_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnouncementCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnouncementOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/members:
    get:
      tags:
      - admin
      summary: List Admin Members
      operationId: list_admin_members_api_v1_admin_members_get
      security:
      - HTTPBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Match against full name or email
          title: Search
        description: Match against full name or email
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMemberListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/payments:
    get:
      tags:
      - admin
      summary: List Admin Payments
      operationId: list_admin_payments_api_v1_admin_payments_get
      security:
      - HTTPBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Match against member name or email
          title: Search
        description: Match against member name or email
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPaymentListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AdminDashboardOut:
      properties:
        stats:
          $ref: '#/components/schemas/AdminStatsOut'
        cash_flow:
          items:
            $ref: '#/components/schemas/RevenueSourceOut'
          type: array
          title: Cash Flow
        budget:
          items:
            $ref: '#/components/schemas/BudgetCategoryOut'
          type: array
          title: Budget
        seat_status:
          $ref: '#/components/schemas/SeatStatusOut'
        seat_occupancy:
          items:
            $ref: '#/components/schemas/SeatOccupancySlotOut'
          type: array
          title: Seat Occupancy
      type: object
      required:
      - stats
      - cash_flow
      - budget
      - seat_status
      - seat_occupancy
      title: AdminDashboardOut
    AdminMemberListOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AdminMemberOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: AdminMemberListOut
    AdminMemberOut:
      properties:
        id:
          type: string
          title: Id
        full_name:
          type: string
          title: Full Name
        email:
          type: string
          title: Email
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
        joined_at:
          type: string
          format: date-time
          title: Joined At
        last_payment_amount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Last Payment Amount
        last_payment_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Payment Label
        last_payment_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Payment At
        plan_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Plan Label
        plan_expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Plan Expires At
        plan_is_active:
          type: boolean
          title: Plan Is Active
        books_reading:
          type: integer
          title: Books Reading
        books_completed:
          type: integer
          title: Books Completed
        reported:
          type: boolean
          title: Reported
        event_registrations:
          type: integer
          title: Event Registrations
      type: object
      required:
      - id
      - full_name
      - email
      - role
      - is_active
      - joined_at
      - last_payment_amount
      - last_payment_label
      - last_payment_at
      - plan_label
      - plan_expires_at
      - plan_is_active
      - books_reading
      - books_completed
      - reported
      - event_registrations
      title: AdminMemberOut
    AdminPaymentListOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AdminPaymentOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: AdminPaymentListOut
    AdminPaymentOut:
      properties:
        id:
          type: string
          title: Id
        member_id:
          type: string
          title: Member Id
        member_name:
          type: string
          title: Member Name
        member_email:
          type: string
          title: Member Email
        amount:
          type: integer
          title: Amount
        label:
          type: string
          title: Label
        status:
          type: string
          title: Status
        plan_months:
          anyOf:
          - type: integer
          - type: 'null'
          title: Plan Months
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - member_id
      - member_name
      - member_email
      - amount
      - label
      - status
      - plan_months
      - created_at
      title: AdminPaymentOut
    AdminStatsOut:
      properties:
        revenue_mtd:
          type: integer
          title: Revenue Mtd
        revenue_trend:
          $ref: '#/components/schemas/TrendOut'
        expenses_mtd:
          type: integer
          title: Expenses Mtd
        expenses_trend:
          $ref: '#/components/schemas/TrendOut'
        net_profit_mtd:
          type: integer
          title: Net Profit Mtd
        net_profit_trend:
          $ref: '#/components/schemas/TrendOut'
        total_members:
          type: integer
          title: Total Members
        total_members_trend:
          $ref: '#/components/schemas/TrendOut'
      type: object
      required:
      - revenue_mtd
      - revenue_trend
      - expenses_mtd
      - expenses_trend
      - net_profit_mtd
      - net_profit_trend
      - total_members
      - total_members_trend
      title: AdminStatsOut
    AnnouncementCreate:
      properties:
        message:
          type: string
          maxLength: 500
          minLength: 1
          title: Message
      type: object
      required:
      - message
      title: AnnouncementCreate
    AnnouncementOut:
      properties:
        recipient_count:
          type: integer
          title: Recipient Count
      type: object
      required:
      - recipient_count
      title: AnnouncementOut
...
    AuditLogEntryOut:
      properties:
        id:
          type: string
          title: Id
        actor_id:
          type: string
          title: Actor Id
        actor_name:
          type: string
          title: Actor Name
        actor_role:
          type: string
          title: Actor Role
        action:
          type: string
          title: Action
        params:
          additionalProperties: true
          type: object
          title: Params
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - actor_id
      - actor_name
      - actor_role
      - action
      - params
      - created_at
      title: AuditLogEntryOut
...
    BudgetCategoryOut:
      properties:
        category:
          $ref: '#/components/schemas/ExpenseCategory'
        budgeted:
          type: integer
          title: Budgeted
        spent:
          type: integer
          title: Spent
      type: object
      required:
      - category
      - budgeted
      - spent
      title: BudgetCategoryOut
...
    ExpenseBreakdownItemOut:
      properties:
        category:
          $ref: '#/components/schemas/ExpenseCategory'
        amount:
          type: integer
          title: Amount
        percent:
          type: number
          title: Percent
      type: object
      required:
      - category
      - amount
      - percent
      title: ExpenseBreakdownItemOut
    ExpenseBreakdownOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ExpenseBreakdownItemOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: ExpenseBreakdownOut
    ExpenseCategory:
      type: string
      enum:
      - staffSalaries
      - bookProcurement
      - utilities
      - marketing
      title: ExpenseCategory
    ExpenseCreate:
      properties:
        category:
          $ref: '#/components/schemas/ExpenseCategory'
        amount:
          type: integer
          exclusiveMinimum: 0.0
          title: Amount
      type: object
      required:
      - category
      - amount
      title: ExpenseCreate
    ExpenseOut:
      properties:
        id:
          type: string
          title: Id
        category:
          type: string
          title: Category
        amount:
          type: integer
          title: Amount
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - category
      - amount
      - created_at
      title: ExpenseOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MemberCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 8
          title: Password
        full_name:
          type: string
          maxLength: 150
          minLength: 1
          title: Full Name
        phone:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Phone
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        role_name:
          type: string
          title: Role Name
          description: Defaults to 'member'
          default: member
      type: object
      required:
      - email
      - password
      - full_name
      title: MemberCreate
    MemberListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MemberOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: MemberListResponse
    MemberOut:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        full_name:
          type: string
          title: Full Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        role:
          $ref: '#/components/schemas/RoleOut'
        is_active:
          type: boolean
          title: Is Active
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - full_name
      - phone
      - avatar_url
      - role
      - is_active
      - last_login_at
      - created_at
      - updated_at
      title: MemberOut
...
    MembershipGrowthMonthOut:
      properties:
        month:
          type: string
          title: Month
        new_members:
          type: integer
          title: New Members
        total_members:
          type: integer
          title: Total Members
      type: object
      required:
      - month
      - new_members
      - total_members
      title: MembershipGrowthMonthOut
    MembershipGrowthOut:
      properties:
        months:
          items:
            $ref: '#/components/schemas/MembershipGrowthMonthOut'
          type: array
          title: Months
      type: object
      required:
      - months
      title: MembershipGrowthOut
...
    MonthlyFigureOut:
      properties:
        month:
          type: string
          title: Month
        revenue:
          type: integer
          title: Revenue
        expenses:
          type: integer
          title: Expenses
        net_profit:
          type: integer
          title: Net Profit
      type: object
      required:
      - month
      - revenue
      - expenses
      - net_profit
      title: MonthlyFigureOut
...
    PaymentCreate:
      properties:
        amount:
          type: integer
          exclusiveMinimum: 0.0
          title: Amount
        label:
          type: string
          maxLength: 255
          minLength: 1
          title: Label
        plan_months:
          anyOf:
          - type: integer
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Plan Months
        coupon_code:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 1
          - type: 'null'
          title: Coupon Code
      type: object
      required:
      - amount
      - label
      title: PaymentCreate
    PaymentOut:
      properties:
        id:
          type: string
          title: Id
        amount:
          type: integer
          title: Amount
        label:
          type: string
          title: Label
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - amount
      - label
      - status
      - created_at
      title: PaymentOut
...
    ProfitAndLossOut:
      properties:
        months:
          items:
            $ref: '#/components/schemas/MonthlyFigureOut'
          type: array
          title: Months
        total_revenue:
          type: integer
          title: Total Revenue
        total_expenses:
          type: integer
          title: Total Expenses
        total_net_profit:
          type: integer
          title: Total Net Profit
      type: object
      required:
      - months
      - total_revenue
      - total_expenses
      - total_net_profit
      title: ProfitAndLossOut
...
    RevenueByPlanItemOut:
      properties:
        label:
          type: string
          title: Label
        amount:
          type: integer
          title: Amount
        count:
          type: integer
          title: Count
      type: object
      required:
      - label
      - amount
      - count
      title: RevenueByPlanItemOut
    RevenueByPlanOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RevenueByPlanItemOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: RevenueByPlanOut
    RevenueSourceOut:
      properties:
        source:
          type: string
          title: Source
        amount:
          type: integer
          title: Amount
      type: object
      required:
      - source
      - amount
      title: RevenueSourceOut
...
    RoleOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: RoleOut
...
    SeatOccupancySlotOut:
      properties:
        hour:
          type: integer
          title: Hour
        percent_filled:
          type: integer
          title: Percent Filled
      type: object
      required:
      - hour
      - percent_filled
      title: SeatOccupancySlotOut
...
    SeatStatusOut:
      properties:
        available:
          type: integer
          title: Available
        booked:
          type: integer
          title: Booked
        total:
          type: integer
          title: Total
      type: object
      required:
      - available
      - booked
      - total
      title: SeatStatusOut
...
    TrendOut:
      properties:
        direction:
          type: string
          title: Direction
        percent:
          type: integer
          title: Percent
      type: object
      required:
      - direction
      - percent
      title: TrendOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer